Revert "Revert "OCPBUGS-55277: backport 4.18""
