+++ pkg/monitortests/kubeapiserver/auditloganalyzer/handle_excessive_applies.go
@@ -19,7 +19,7 @@ func CheckForExcessiveApplies() *excessiveApplies {
 	}
 }
 
-func (s *excessiveApplies) HandleAuditLogEvent(auditEvent *auditv1.Event, beginning, end *metav1.MicroTime) {
+func (s *excessiveApplies) HandleAuditLogEvent(auditEvent *auditv1.Event, beginning, end *metav1.MicroTime, nodeName string) {
 	if beginning != nil && auditEvent.RequestReceivedTimestamp.Before(beginning) || end != nil && end.Before(&auditEvent.RequestReceivedTimestamp) {
 		return
 	}
