+++ pkg/monitortests/kubeapiserver/auditloganalyzer/handle_invalid_requests.go
@@ -86,7 +86,7 @@ func isApply(auditEvent *auditv1.Event) bool {
 	return true
 }
 
-func (s *invalidRequests) HandleAuditLogEvent(auditEvent *auditv1.Event, beginning, end *metav1.MicroTime) {
+func (s *invalidRequests) HandleAuditLogEvent(auditEvent *auditv1.Event, beginning, end *metav1.MicroTime, nodeName string) {
 	if beginning != nil && auditEvent.RequestReceivedTimestamp.Before(beginning) || end != nil && end.Before(&auditEvent.RequestReceivedTimestamp) {
 		return
 	}
